requirements.txt
beautifulsoup4>=4.11.1
pytest>=7.1.2
python-dateutil>=2.8.2
requests>=2.27.1
selenium>=4.1.3
